# apps/cloud-server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin-api/industries": {
      "post": {
        "operationId": "IndustryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIndustryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Industry created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed."
          },
          "401": {
            "description": "Unauthorized."
          },
          "409": {
            "description": "Industry with this code or slug already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new industry",
        "tags": [
          "Admin - Industries"
        ]
      }
    },
    "/admin-api/industries/table": {
      "get": {
        "operationId": "IndustryController_findForTable",
        "parameters": [
          {
            "name": "searchColumn",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchValue",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Industries retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustryTableResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List industries for data table (server-stored state)",
        "tags": [
          "Admin - Industries"
        ]
      }
    },
    "/admin-api/industries/select": {
      "get": {
        "operationId": "IndustryController_findForSelect",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter by label",
            "schema": {
              "example": "united",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of results",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of results to skip",
            "schema": {
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "values",
            "required": false,
            "in": "query",
            "description": "Comma-separated values to fetch specific options",
            "schema": {
              "example": "1,2,3",
              "type": "string"
            }
          },
          {
            "name": "excludeIds",
            "required": false,
            "in": "query",
            "description": "Comma-separated IDs to exclude from results (already selected)",
            "schema": {
              "example": "5,10",
              "type": "string"
            }
          },
          {
            "name": "valueKey",
            "required": false,
            "in": "query",
            "description": "Column name for option value",
            "schema": {
              "default": "id",
              "example": "id",
              "type": "string"
            }
          },
          {
            "name": "labelKey",
            "required": false,
            "in": "query",
            "description": "Column name for option label",
            "schema": {
              "default": "name",
              "example": "name",
              "type": "string"
            }
          },
          {
            "name": "descriptionKey",
            "required": false,
            "in": "query",
            "description": "Column name for option description",
            "schema": {
              "example": "description",
              "type": "string"
            }
          },
          {
            "name": "groupIdKey",
            "required": false,
            "in": "query",
            "description": "Column name for group ID",
            "schema": {
              "example": "regionId",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Industry options retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustrySelectResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get industries for select component",
        "tags": [
          "Admin - Industries"
        ]
      }
    },
    "/admin-api/industries/{id}": {
      "patch": {
        "operationId": "IndustryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Industry UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIndustryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Industry updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Industry not found."
          },
          "409": {
            "description": "Industry with this code or slug already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an industry",
        "tags": [
          "Admin - Industries"
        ]
      },
      "delete": {
        "operationId": "IndustryController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Industry UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Industry deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Industry not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an industry",
        "tags": [
          "Admin - Industries"
        ]
      }
    },
...
    "/admin-api/prices": {
      "post": {
        "operationId": "PriceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePriceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Price created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or invalid reference ID."
          },
          "401": {
            "description": "Unauthorized."
          },
          "409": {
            "description": "Price for this combination already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new price entry",
        "tags": [
          "Admin - Prices"
        ]
      }
    },
    "/admin-api/prices/plan/{planId}/table": {
      "get": {
        "operationId": "PriceController_findForTable",
        "parameters": [
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "description": "Plan UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prices retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricesTableResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List prices for a plan in data table format",
        "tags": [
          "Admin - Prices"
        ]
      }
    },
    "/admin-api/prices/{id}": {
      "patch": {
        "operationId": "PriceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Price UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePriceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Price updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or invalid reference ID."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Price not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a price",
        "tags": [
          "Admin - Prices"
        ]
      },
      "delete": {
        "operationId": "PriceController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Price UUID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Price deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Price not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a price",
        "tags": [
          "Admin - Prices"
        ]
      }
    },
...
    "/cloud-api/industries/select": {
      "get": {
        "description": "Returns paginated industry options for quantum-ui Select.",
        "operationId": "IndustryController_findForSelect",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter by label",
            "schema": {
              "example": "united",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of results",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of results to skip",
            "schema": {
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "values",
            "required": false,
            "in": "query",
            "description": "Comma-separated values to fetch specific options",
            "schema": {
              "example": "1,2,3",
              "type": "string"
            }
          },
          {
            "name": "excludeIds",
            "required": false,
            "in": "query",
            "description": "Comma-separated IDs to exclude from results (already selected)",
            "schema": {
              "example": "5,10",
              "type": "string"
            }
          },
          {
            "name": "valueKey",
            "required": false,
            "in": "query",
            "description": "Column name for option value",
            "schema": {
              "default": "id",
              "example": "id",
              "type": "string"
            }
          },
          {
            "name": "labelKey",
            "required": false,
            "in": "query",
            "description": "Column name for option label",
            "schema": {
              "default": "name",
              "example": "name",
              "type": "string"
            }
          },
          {
            "name": "descriptionKey",
            "required": false,
            "in": "query",
            "description": "Column name for option description",
            "schema": {
              "example": "description",
              "type": "string"
            }
          },
          {
            "name": "groupIdKey",
            "required": false,
            "in": "query",
            "description": "Column name for group ID",
            "schema": {
              "example": "regionId",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Industry options retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustrySelectResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get industries for select component",
        "tags": [
          "Industries"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vritti Cloud API",
    "description": "Internal API for Vritti SaaS Platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://local.vrittiai.com:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "CreateIndustryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the industry",
            "example": "Healthcare"
          },
          "code": {
            "type": "string",
            "description": "Unique code identifier for the industry",
            "example": "healthcare"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the industry",
            "example": "Healthcare and medical services"
          }
        },
        "required": [
          "name",
          "code"
        ]
      },
      "IndustryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "IndustryTableResponseDto": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndustryDto"
            }
          },
          "count": {
            "type": "number"
          },
          "state": {
            "type": "object",
            "description": "Current active filter/sort/visibility state"
          },
          "activeViewId": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "result",
          "count",
          "state"
        ]
      },
      "IndustrySelectOptionDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "label": {
            "type": "string",
            "example": "Technology"
          }
        },
        "required": [
          "value",
          "label"
        ]
      },
      "IndustrySelectResponseDto": {
        "type": "object",
        "properties": {
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndustrySelectOptionDto"
            }
          },
          "hasMore": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "options",
          "hasMore"
        ]
      },
      "UpdateIndustryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the industry",
            "example": "Healthcare"
          },
          "code": {
            "type": "string",
            "description": "Unique code identifier for the industry",
            "example": "healthcare"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the industry",
            "example": "Healthcare and medical services"
          }
        }
      },
...
      "CreatePriceDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string",
            "description": "Plan UUID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "industryId": {
            "type": "string",
            "description": "Industry UUID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "regionId": {
            "type": "string",
            "description": "Region UUID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "providerId": {
            "type": "string",
            "description": "Provider UUID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "price": {
            "type": "string",
            "description": "Price amount",
            "example": "99.99"
          },
          "currency": {
            "type": "string",
            "description": "ISO 4217 currency code",
            "example": "INR"
          }
        },
        "required": [
          "planId",
          "industryId",
          "regionId",
          "providerId",
          "price",
          "currency"
        ]
      },
      "PriceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "planId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "industryId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "regionId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "providerId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "price": {
            "type": "string",
            "example": "99.99"
          },
          "currency": {
            "type": "string",
            "example": "INR"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "planId",
          "industryId",
          "regionId",
          "providerId",
          "price",
          "currency",
          "createdAt"
        ]
      },
      "PriceDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "planId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "industryId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "regionId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "regionName": {
            "type": "string",
            "example": "Mumbai"
          },
          "regionCode": {
            "type": "string",
            "example": "ap-south-1"
          },
          "providerId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "providerName": {
            "type": "string",
            "example": "Amazon Web Services"
          },
          "providerCode": {
            "type": "string",
            "example": "AWS"
          },
          "price": {
            "type": "string",
            "example": "99.99"
          },
          "currency": {
            "type": "string",
            "example": "INR"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "planId",
          "industryId",
          "regionId",
          "regionName",
          "regionCode",
          "providerId",
          "providerName",
          "providerCode",
          "price",
          "currency",
          "createdAt"
        ]
      },
      "PricesTableResponseDto": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceDetailDto"
            }
          },
          "count": {
            "type": "number"
          },
          "state": {
            "type": "object",
            "description": "Current active filter/sort/visibility state"
          },
          "activeViewId": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "result",
          "count",
          "state"
        ]
      },
      "UpdatePriceDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string",
            "description": "Plan UUID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "industryId": {
            "type": "string",
            "description": "Industry UUID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "regionId": {
            "type": "string",
            "description": "Region UUID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "providerId": {
            "type": "string",
            "description": "Provider UUID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "price": {
            "type": "string",
            "description": "Price amount",
            "example": "99.99"
          },
          "currency": {
            "type": "string",
            "description": "ISO 4217 currency code",
            "example": "INR"
          }
        }
      },
...
    }
  }
}